requirements-test.txt
pytest>=6.2.5
pytest-cov>=2.12.1
pytest-mock>=3.6.1
pyyaml>=5.4.1